extract datalayer rendering into public service
