+++ Services/DatalayerRendererInterface.php
@@ -1,0 +1,31 @@
+<?php
+/**
+ * Tag Manager
+ * Copyright (c) Webmatch GmbH
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+namespace WbmTagManager\Services;
+
+/**
+ * Interface DatalayerRendererInterface
+ */
+interface DatalayerRendererInterface
+{
+    /**
+     * @param $moduleName string
+     * @param $context array
+     *
+     * @return array
+     */
+    public function render($moduleName, $context);
+}
